FIX : 자동로그인 로직변경
